% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icbs.R
\name{make_connector}
\alias{make_connector}
\title{Make a connector panel}
\usage{
make_connector(
  direction = "split",
  sign = "+",
  line_col = "black",
  line_size = 0.5,
  label_size = 10,
  label_fill = "white"
)
}
\arguments{
\item{direction}{character, either "split" (left-to-right branching) or
"merge" (right-to-left converging). Default: "split".}

\item{sign}{character label to display at the center of the connector.
Default: "+".}

\item{line_col}{color of the connector lines. Default: "black".}

\item{line_size}{linewidth of the connector lines. Default: 0.5.}

\item{label_size}{text size of the center label. Default: 10.}

\item{label_fill}{background fill of the center label. Default: "white".}
}
\value{
ggplot
}
\description{
Creates a standalone connector ggplot element for use in multi-panel layouts.
Supports split (one-to-two) and merge (two-to-one) directions.
}
