% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stwd.R
\name{one_to_split_to_two_layout}
\alias{one_to_split_to_two_layout}
\title{make a 1 split 2 plot}
\usage{
one_to_split_to_two_layout(fig, first_level_sign = "+")
}
\arguments{
\item{fig}{list of plots}

\item{first_level_sign}{sign}
}
\value{
ggplot
}
\description{
make a 1 split 2 plot
}
