% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icbs.R
\name{scale_color_ibcs_scenario}
\alias{scale_color_ibcs_scenario}
\title{IBCS Scenario Color Scale}
\usage{
scale_color_ibcs_scenario(...)
}
\arguments{
\item{...}{Additional arguments passed to discrete_scale}
}
\value{
A ggplot2 scale object
}
\description{
A color scale for IBCS scenario notation. Used for bar outlines,
especially for Plan (PL) which shows as outlined bars.
}
