% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icbs.R
\name{theme_marquee}
\alias{theme_marquee}
\title{Apply Marquee Styling to Common Text Elements}
\usage{
theme_marquee()
}
\value{
A ggplot2 theme object.
}
\description{
A theme addition that enables marquee rendering for title, subtitle,
caption, and legend text. Use after your base theme.
}
\examples{
\dontrun{
ggplot(...) +
    theme_ibcs() +
    theme_marquee()
}

}
