% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icbs.R
\name{two_to_merge_to_one_layout}
\alias{two_to_merge_to_one_layout}
\title{Make a 2 merge to 1 plot}
\usage{
two_to_merge_to_one_layout(fig, first_level_sign = "+")
}
\arguments{
\item{fig}{list of 3 ggplot objects: two left panels and one right panel}

\item{first_level_sign}{sign to display on the connector (default: "+")}
}
\value{
ggplot
}
\description{
Creates a layout where two plots on the left converge via a connector
into a single plot on the right.
}
